+++ src/Crash.Common/Events/CrashInitArgs.cs
@@ -2,20 +2,18 @@
 
 namespace Crash.Common.Events
 {
-	/// <summary>
-	///     Captures the initialization of <see cref="Change" />s
-	/// </summary>
+
 	public sealed class CrashInitArgs : CrashEventArgs
 	{
-		/// <summary>
-		///     The updated <see cref="Change" />s
-		/// </summary>
-		public readonly IEnumerable<Change> Changes;
 
-		public CrashInitArgs(CrashDoc crashDoc, IEnumerable<Change> changes)
+		public int ChangeCount { get; }
+
+		public CrashInitArgs(CrashDoc crashDoc, int changeCount)
 			: base(crashDoc)
 		{
-			Changes = changes;
+			ChangeCount = changeCount;
 		}
+
 	}
+
 }
